+++ components/SEO.tsx
@@ -14,7 +14,6 @@ const SEO: React.FC<SEOProps> = ({ title, description, image }) => {
       <meta property="og:title" content={title} />
       <meta property="og:description" content={description} />
       <meta property="og:image" content={image} />
-      {/*<meta name="twitter:card" content="summary_large_image" />*/}
       <meta property="twitter:domain" content="luddy.dance" />
       <meta property="twitter:url" content="https://luddy.dance" />
       <meta name="twitter:title" content={title} />
